TensorsLite

Documentation for TensorsLite.

TensorsLite.AbstractTensor

TensorsLite.Tensor

TensorsLite._non_StaticBool_type

TensorsLite.nonzero_eltype

TensorsLite.tensor_type_1Dx

TensorsLite.tensor_type_1Dy

TensorsLite.tensor_type_1Dz

TensorsLite.tensor_type_2Dxy

TensorsLite.tensor_type_2Dxz

TensorsLite.tensor_type_2Dyz

TensorsLite.tensor_type_3D

TensorsLite.AbstractTensor — Type

AbstractTensor{T, N} <: AbstractArray{T, N}

Supertype of all Tensor types. Represents any Nth order tensor with eltype T.

TensorsLite.Tensor — Type

Tensor{T, N, Tx, Ty, Tz} <: AbstractTensor{T, N}

A Nth order tensor with eltype T. Higher order tensors are implemented as vectors of lower order tensors.

TensorsLite._non_StaticBool_type — Method

_non_StaticBool_type(::Type{Union{T,Zero,One}) -> T
_non_StaticBool_type(::Type{Union{T,Zero}) -> T
_non_StaticBool_type(::Type{Union{T,One}) -> T
_non_StaticBool_type(::Type{Union{Zero,One}) -> One

TensorsLite.nonzero_eltype — Method

nonzero_eltype(::Type{AbstractArray{Union{T,Zero}}) -> T
nonzero_eltype(::Type{AbstractArray{Union{T,Zero,One}}) -> T
nonzero_eltype(::Type{AbstractArray{Union{T,One}}) -> T
nonzero_eltype(::Type{AbstractArray{Union{Zero,One}}) -> One

TensorsLite.tensor_type_1Dx — Method

tensor_type_1Dx(::Val{N}, T::DataType) -> Type{Tensor{Union{T,Zero},N,Tx,Ty,Tz}} where {Tx,Ty,Tz}

Return the type of a 1D in x Nth order tensor with element type of T.

TensorsLite.tensor_type_1Dy — Method

tensor_type_1Dy(::Val{N}, T::DataType) -> Type{Tensor{Union{T,Zero},N,Tx,Ty,Tz}} where {Tx,Ty,Tz}

Return the type of a 1D in y Nth order tensor with element type of T.

TensorsLite.tensor_type_1Dz — Method

tensor_type_1Dz(::Val{N}, T::DataType) -> Type{Tensor{Union{T,Zero},N,Tx,Ty,Tz}} where {Tx,Ty,Tz}

Return the type of a 1D in z Nth order tensor with element type of T.

TensorsLite.tensor_type_2Dxy — Method

tensor_type_2Dxy(::Val{N}, T::DataType) -> Type{Tensor{Union{T,Zero},N,Tx,Ty,Tz}} where {Tx,Ty,Tz}

Return the type of a 2D in xy Nth order tensor with element type of T.

TensorsLite.tensor_type_2Dxz — Method

tensor_type_2Dxz(::Val{N}, T::DataType) -> Type{Tensor{Union{T,Zero},N,Tx,Ty,Tz}} where {Tx,Ty,Tz}

Return the type of a 2D in xz Nth order tensor with element type of T.

TensorsLite.tensor_type_2Dyz — Method

tensor_type_2Dyz(::Val{N}, T::DataType) -> Type{Tensor{Union{T,Zero},N,Tx,Ty,Tz}} where {Tx,Ty,Tz}

Return the type of a 2D in yz Nth order tensor with element type of T.

TensorsLite.tensor_type_3D — Method

tensor_type_3D(::Val{N}, T::DataType) -> Type{Tensor{T,N,Tx,Ty,Tz}} where {Tx,Ty,Tz}

Return the type of a 3D Nth order tensor with element type of T.